Moved some components to UI for organization
