+++ components/PDPComponent/AddAndRemoveProduct.tsx
@@ -3,11 +3,11 @@ import { useMutation, useQuery } from '@apollo/react-hooks';
 import {
   CircularProgress, Typography, DialogContent, Dialog, ButtonGroup, Button,
 } from '@material-ui/core';
-import { ProductsType, PRODUCTS_QUERY } from '../ProductsGrid/_types';
+import { ProductsType, PRODUCTS_QUERY } from '../ui/ProductsGrid/_types';
 import { CHECKOUT_LINE_ITEMS_REPLACE_MUTATION, GET_CART_QUERY, GetCartRequest } from '../CartController/_types';
 import { getLineItems } from './_types';
-import { Styled as StyledThumbnail } from '../ProductThumbnail/_styles';
-import { Styled as StyledGrid } from '../ProductsGrid/_styles';
+import { Styled as StyledThumbnail } from '../ui/ProductThumbnail/_styles';
+import { Styled as StyledGrid } from '../ui/ProductsGrid/_styles';
 
 type Props = {
   variantId: string;
