+++ components/sections/PageContentSection/index.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from 'react';
 import { Styled } from './_styles';
 import { TestInterface } from './_types';
-import CTAButton from '../../CTAButton';
+import CTAButton from '../../ui/CTAButton';
 import { theme } from '../../../views/theme';
 
 type Props = {
