+++ components/ui/NavBar/NavBar.tsx
@@ -17,11 +17,11 @@ import {
   COLLECTIONS_QUERY, Collections,
 } from './_types';
 import './_style.scss';
-import CartData from '../CartContent/CartData';
-import CartController from '../CartController/index';
+import CartData from '../../CartContent/CartData';
+import CartController from '../../CartController';
 import NavItems from './navItems.json';
 import NavBarItem, { ProductItem } from '../NavBarItem';
-import AccountPopup from '../AccountPopup';
+import AccountPopup from '../../AccountPopup';
 
 const HomeLink = styled(Typography)`
 &&& {
