+++ components/ui/ProductThumbnail/_styles.ts
