+++ components/ui/ProductThumbnail/_types.ts
