+++ components/ui/ProductThumbnail/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import Link from 'next/link';
-import { ProductPriceRange } from '../../models';
+import { ProductPriceRange } from '../../../models';
 import { Styled } from './_styles';
-import { GET_CART_QUERY } from '../CartController/_types';
+import { GET_CART_QUERY } from '../../CartController/_types';
 import { GetCartResponse, LineItems, LineItemsInput } from './_types';
-import AddToCart from '../PDPComponent/AddAndRemoveProduct';
+import AddToCart from '../../PDPComponent/AddAndRemoveProduct';
 
 type Props = {
   id: string;
