+++ pages/cart.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Head from 'next/head';
 import { Typography, withTheme } from '@material-ui/core';
 import withData from '../lib/apollo';
-import NavBar from '../components/NavBar/NavBar';
+import NavBar from '../components/ui/NavBar/NavBar';
 import CartData from '../components/CartContent/CartData';
 import PageContainer, { PageSize } from '../views/layouts/PageContainer';
 import { withMuiApp } from '../hocs/withMui';
