+++ pages/index.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import Head from 'next/head';
 import { withTheme } from '@material-ui/core';
 import { useRouter } from 'next/router';
-import ProductsGrid from '../components/ProductsGrid';
+import ProductsGrid from '../components/ui/ProductsGrid';
 import { ProductSortKeys } from '../models';
 import withData from '../lib/apollo';
 import PageContainer, { PageSize } from '../views/layouts/PageContainer';
-import NavBar from '../components/NavBar/NavBar';
+import NavBar from '../components/ui/NavBar/NavBar';
 import HeroBanner from '../components/sections/HeroBanner';
 import { withMuiApp } from '../hocs/withMui';
 import FooterNav from '../components/FooterNav';
