+++ pages/login.tsx
@@ -5,7 +5,7 @@ import { ThemeProvider } from '@material-ui/styles';
 import { withTheme } from '@material-ui/core';
 import Head from 'next/head';
 import withData from '../lib/apollo';
-import NavBar from '../components/NavBar/NavBar';
+import NavBar from '../components/ui/NavBar/NavBar';
 import { theme } from '../views/theme';
 import { withMuiApp } from '../hocs/withMui';
 import LoginForm from '../components/LoginForm';
