+++ pages/product.tsx
@@ -4,7 +4,7 @@ import { ParsedUrlQuery } from 'querystring';
 import { ThemeProvider } from '@material-ui/styles';
 import { withTheme } from '@material-ui/core';
 import withData from '../lib/apollo';
-import NavBar from '../components/NavBar/NavBar';
+import NavBar from '../components/ui/NavBar/NavBar';
 import PDPComponent from '../components/PDPComponent';
 import { theme } from '../views/theme';
 import { withMuiApp } from '../hocs/withMui';
